refactor(test): follow project test conventions for session resolution tests
- Move database helpers to api-test-helpers.ts (allowed by lint rules)
- Use testContext().createAgentCompose() instead of direct DB operations
- Add createTestSessionWithConversation helper for full data chain setup
- Replace deleted compose test with session-has-no-conversation test
