+++ turbo/apps/web/src/lib/slack/handlers/__tests__/session-resolution.test.ts
@@ -1,136 +1,26 @@
 import { describe, it, expect, beforeEach } from "vitest";
 import { testContext, uniqueId } from "../../../../__tests__/test-helpers";
-import { resolveSessionCompose, getWorkspaceAgent } from "../shared";
-import { agentSessions } from "../../../../db/schema/agent-session";
 import {
-  agentComposes,
-  agentComposeVersions,
-} from "../../../../db/schema/agent-compose";
-import { agentRuns } from "../../../../db/schema/agent-run";
-import { conversations } from "../../../../db/schema/conversation";
-import { scopes } from "../../../../db/schema/scope";
-import { eq } from "drizzle-orm";
-import { initServices } from "../../../init-services";
+  createTestAgentSession,
+  createTestSessionWithConversation,
+} from "../../../../__tests__/api-test-helpers";
+import { resolveSessionCompose, getWorkspaceAgent } from "../shared";
 
 const context = testContext();
 
 describe("resolveSessionCompose", () => {
   beforeEach(() => {
     context.setupMocks();
-    initServices();
   });
-
-  async function createTestScope(userId: string) {
-    const [scope] = await globalThis.services.db
-      .insert(scopes)
-      .values({
-        slug: uniqueId("scope"),
-        type: "personal",
-        ownerId: userId,
-      })
-      .returning();
-    return scope!;
-  }
-
-  async function createTestCompose(
-    userId: string,
-    scopeId: string,
-    name: string,
-  ) {
-    const [compose] = await globalThis.services.db
-      .insert(agentComposes)
-      .values({
-        userId,
-        scopeId,
-        name,
-      })
-      .returning();
-    return compose!;
-  }
-
-  async function createTestComposeVersion(composeId: string, userId: string) {
-    const versionId = uniqueId("version");
-    await globalThis.services.db.insert(agentComposeVersions).values({
-      id: versionId,
-      composeId,
-      content: { name: "test-agent", model: "claude-3-5-sonnet-20241022" },
-      createdBy: userId,
-    });
-    // Update compose to point to this version
-    await globalThis.services.db
-      .update(agentComposes)
-      .set({ headVersionId: versionId })
-      .where(eq(agentComposes.id, composeId));
-    return versionId;
-  }
-
-  async function createTestRun(userId: string, versionId: string) {
-    const [run] = await globalThis.services.db
-      .insert(agentRuns)
-      .values({
-        userId,
-        agentComposeVersionId: versionId,
-        status: "completed",
-        prompt: "test prompt",
-      })
-      .returning({ id: agentRuns.id });
-    return run!;
-  }
-
-  async function createTestConversation(runId: string) {
-    const [conversation] = await globalThis.services.db
-      .insert(conversations)
-      .values({
-        runId,
-        cliAgentType: "claude",
-        cliAgentSessionId: uniqueId("cli-session"),
-        cliAgentSessionHistory: "[]",
-      })
-      .returning({ id: conversations.id });
-    return conversation!;
-  }
-
-  async function createTestSession(userId: string, agentComposeId: string) {
-    const [session] = await globalThis.services.db
-      .insert(agentSessions)
-      .values({ userId, agentComposeId })
-      .returning({ id: agentSessions.id });
-    return session!;
-  }
-
-  async function createTestSessionWithConversation(
-    userId: string,
-    agentComposeId: string,
-  ) {
-    // Create compose version
-    const versionId = await createTestComposeVersion(agentComposeId, userId);
-    // Create run
-    const run = await createTestRun(userId, versionId);
-    // Create conversation
-    const conversation = await createTestConversation(run.id);
-    // Create session with conversation
-    const [session] = await globalThis.services.db
-      .insert(agentSessions)
-      .values({
-        userId,
-        agentComposeId,
-        conversationId: conversation.id,
-      })
-      .returning({ id: agentSessions.id });
-    return session!;
-  }
 
   describe("when session exists and belongs to user", () => {
     it("should return compose info from session", async () => {
       const userId = uniqueId("test-user");
 
-      // Create scope and compose
-      const scope = await createTestScope(userId);
-      const compose = await createTestCompose(
-        userId,
-        scope.id,
-        "session-agent",
-      );
+      // Create compose using testContext helper
+      const compose = await context.createAgentCompose(userId, {
+        name: "session-agent",
+      });
 
       // Create an agent session with conversation (required by validateAgentSession)
       const session = await createTestSessionWithConversation(
@@ -165,16 +55,13 @@ describe("resolveSessionCompose", () => {
       const ownerUserId = uniqueId("owner-user");
       const otherUserId = uniqueId("other-user");
 
-      // Create scope and compose for owner
-      const scope = await createTestScope(ownerUserId);
-      const compose = await createTestCompose(
-        ownerUserId,
-        scope.id,
-        "owner-agent",
-      );
+      // Create compose for owner
+      const compose = await context.createAgentCompose(ownerUserId, {
+        name: "owner-agent",
+      });
 
-      // Create a session owned by the owner
-      const session = await createTestSession(ownerUserId, compose.id);
+      // Create a session owned by the owner (without conversation - auth check happens first)
+      const session = await createTestAgentSession(ownerUserId, compose.id);
 
       // Try to resolve with other user - should fail authorization
       const result = await resolveSessionCompose(session.id, otherUserId);
@@ -183,27 +70,19 @@ describe("resolveSessionCompose", () => {
     });
   });
 
-  describe("when session compose is deleted", () => {
-    it("should return undefined if compose no longer exists", async () => {
+  describe("when session has no conversation", () => {
+    it("should return undefined", async () => {
       const userId = uniqueId("test-user");
 
-      // Create scope and compose
-      const scope = await createTestScope(userId);
-      const compose = await createTestCompose(
-        userId,
-        scope.id,
-        "deleted-agent",
-      );
+      // Create compose
+      const compose = await context.createAgentCompose(userId, {
+        name: "test-agent",
+      });
 
-      // Create a session
-      const session = await createTestSession(userId, compose.id);
+      // Create a session without conversation
+      const session = await createTestAgentSession(userId, compose.id);
 
-      // Delete the compose (simulate orphaned session)
-      await globalThis.services.db
-        .delete(agentComposes)
-        .where(eq(agentComposes.id, compose.id));
-
-      // Resolve should return undefined since compose doesn't exist
+      // Resolve should return undefined since session has no conversation
       const result = await resolveSessionCompose(session.id, userId);
 
       expect(result).toBeUndefined();
@@ -214,35 +93,20 @@ describe("resolveSessionCompose", () => {
 describe("getWorkspaceAgent", () => {
   beforeEach(() => {
     context.setupMocks();
-    initServices();
   });
 
   it("should return agent info for valid composeId", async () => {
     const userId = uniqueId("test-user");
 
-    // Create scope and compose directly
-    const [scope] = await globalThis.services.db
-      .insert(scopes)
-      .values({
-        slug: uniqueId("scope"),
-        type: "personal",
-        ownerId: userId,
-      })
-      .returning();
+    // Create compose using testContext helper
+    const compose = await context.createAgentCompose(userId, {
+      name: "test-agent",
+    });
 
-    const [compose] = await globalThis.services.db
-      .insert(agentComposes)
-      .values({
-        userId,
-        scopeId: scope!.id,
-        name: "test-agent",
-      })
-      .returning();
-
-    const result = await getWorkspaceAgent(compose!.id);
+    const result = await getWorkspaceAgent(compose.id);
 
     expect(result).toBeDefined();
-    expect(result!.id).toBe(compose!.id);
+    expect(result!.id).toBe(compose.id);
     expect(result!.name).toBe("test-agent");
   });
